test_standalone_build/test_standalone_mcp/requirements.txt
google-generativeai>=0.8.5
lancedb>=0.22.0
pydantic>=2.11.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0

# MCP Server Requirements
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
